ss font olmuyor

+++ CCARS/CCARS/UIHelpers/UICompanentsHelper.swift
@@ -8,11 +8,11 @@
 import UIKit
 
 struct UICompanentsHelper{
-    static func creatCustomTextLabel(text : String , size : Int , fontName : String , textColor: UIColor , backgraondColor: UIColor? ,borderColor : CGColor? ,cornerRadius : CGFloat? , borderWidth : CGFloat?) -> UILabel {
+    static func creatCustomTextLabel(text : String , size : CGFloat , fontName : String , textColor: UIColor , backgraondColor: UIColor? ,borderColor : CGColor? ,cornerRadius : CGFloat? , borderWidth : CGFloat?) -> UILabel {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
         label.text = text
-        label.font = UIFont(name: fontName, size: CGFloat(size))
+        label.font = UIFont(name: fontName, size: size)
         label.textColor = textColor
         label.numberOfLines = 0
         label.clipsToBounds = true
